Add more dependencies and plugins in pom.xml

See merge request openclassrooms/projet5!13

+++ src/main/java/com/safetynet/alerts/service/MedicalRecordServiceImpl.java
@@ -117,7 +117,7 @@ public class MedicalRecordServiceImpl implements MedicalRecordService {
         logger.debug("firstName : " + firstName + " | lastName : " + lastName);
 
         for (MedicalRecordDTO medicalRecordDTO : medicalRecordRepository.getMedicalRecordList()) {
-            if (medicalRecordDTO.getFirstName().equals(firstName) & medicalRecordDTO.getLastName().equals(lastName)) {
+            if (medicalRecordDTO.getFirstName().equals(firstName) && medicalRecordDTO.getLastName().equals(lastName)) {
                 logger.debug("medicalRecordDTO : " + medicalRecordDTO.toString());
                 medicalRecordRepository.deleteMedicalRecord(medicalRecordDTO);
                 return true;
